Внес правки по заданию 'Начинаем программировать'
