+++ project/src/components/app/type.ts
@@ -1,0 +1,5 @@
+type AppScreenProps = {
+  offersCount: number;
+}
+
+export type {AppScreenProps};
