+++ project/src/components/city-component/type.ts
@@ -1,0 +1,5 @@
+type CityComponentProps = {
+  city: string;
+}
+
+export type {CityComponentProps};
