+++ project/src/components/main-screen/type.ts
@@ -1,0 +1,5 @@
+type MainScreenProps = {
+  offersCount: number;
+}
+
+export type {MainScreenProps};
